requirements.txt
coverage==6.5.0
dvc==2.28.0
numpy==1.23.1
pandas==1.1.3
scikit-learn==1.1.1
pyyaml==5.3.1